Add Pull to Refresh data
