+++ presentation/src/main/java/com/nuweather/di/ViewModelModule.kt
@@ -1,13 +1,9 @@
 package com.nuweather.di
 
 import com.nuweather.feature.currentweather.CurrentWeatherViewModel
-import com.nuweather.feature.home.HomeViewModel
-import com.nuweather.feature.temperaturechart.TemperatureChartViewModel
 import org.koin.android.viewmodel.ext.koin.viewModel
 import org.koin.dsl.module.module
 
 val viewModelModule = module {
-    viewModel { HomeViewModel() }
-    viewModel { CurrentWeatherViewModel(get(), get()) }
-    viewModel { TemperatureChartViewModel(get(), get()) }
+    viewModel { CurrentWeatherViewModel(get(), get(), get()) }
 }
