+++ presentation/src/main/java/com/nuweather/feature/MainActivity.kt
@@ -3,7 +3,7 @@ package com.nuweather.feature
 import android.os.Bundle
 import com.nuweather.R
 import com.nuweather.base.BaseActivity
-import com.nuweather.feature.home.HomeFragment
+import com.nuweather.feature.currentweather.CurrentWeatherFragment
 
 class MainActivity : BaseActivity() {
 
@@ -13,7 +13,7 @@ class MainActivity : BaseActivity() {
 
         if (savedInstanceState == null) {
             supportFragmentManager.beginTransaction()
-                .replace(R.id.container, HomeFragment.newInstance())
+                .replace(R.id.container, CurrentWeatherFragment.newInstance())
                 .commitNow()
 
         }
