--- presentation/src/main/res/layout/fragment_home.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:bind="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".feature.MainActivity">
-
-    <data>
-
-        <variable
-            name="viewModel"
-            type="com.nuweather.feature.home.HomeViewModel" />
-    </data>
-
-    <android.support.constraint.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-
-        <android.support.v7.widget.AppCompatImageView
-            android:id="@+id/weatherBackground"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:scaleType="centerCrop"
-            android:src="@drawable/bg_morning"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            bind:imageResource="@{viewModel.weatherBackground}" />
-
-        <FrameLayout
-            android:id="@+id/currentWeather"
-            android:layout_width="@dimen/dp_0"
-            android:layout_height="@dimen/dp_0"
-            app:layout_constraintBottom_toTopOf="@+id/temperatureChart"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <FrameLayout
-            android:id="@+id/temperatureChart"
-            android:layout_width="@dimen/dp_0"
-            android:layout_height="@dimen/dp_100"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/currentWeather" />
-
-        <TextView
-            android:id="@+id/goToDailyWeather"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/dp_24"
-            android:layout_marginTop="@dimen/dp_8"
-            android:layout_marginEnd="@dimen/dp_24"
-            android:layout_marginBottom="@dimen/dp_24"
-            android:background="@drawable/bg_rounded_corner_black_20"
-            android:drawableEnd="@drawable/ic_right_arrow"
-            android:fontFamily="sans-serif-light"
-            android:paddingStart="@dimen/dp_16"
-            android:paddingTop="@dimen/dp_8"
-            android:paddingEnd="@dimen/dp_16"
-            android:paddingBottom="@dimen/dp_8"
-            android:text="@string/daily_forecast_16_days"
-            android:textColor="@color/colorWhite"
-            android:textSize="@dimen/sp_18"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/temperatureChart" />
-
-    </android.support.constraint.ConstraintLayout>
-</layout>
