--- presentation/src/main/res/layout/fragment_temperature_chart.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".feature.MainActivity">
-
-    <data>
-
-        <variable
-            name="viewModel"
-            type="com.nuweather.feature.temperaturechart.TemperatureChartViewModel" />
-    </data>
-
-    <android.support.constraint.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        tools:background="@drawable/bg_morning">
-
-    </android.support.constraint.ConstraintLayout>
-
-</layout>
